test-release-docs/requirements.txt
atlassian-python-api
requests
jinja2
